requirements.txt
python-telegram-bot==20.7
playwright==1.43.0
google-generativeai